refactor(server): organize & rename middlewares
