+++ server/middlewares/error-handlers/http-error-handler.ts
