+++ server/middlewares/error-handlers/index.ts
@@ -1,0 +1,2 @@
+export * from './http-error-handler';
+export * from './zod-error-handler';
