+++ server/middlewares/index.ts
@@ -1,5 +1,3 @@
-export * from './http-error-handler';
+export * from './error-handlers';
+export * from './validators';
 export * from './middleware-wrapper';
-export * from './mongo-id-param-validator';
-export * from './zod-error-handler';
-export * from './zod-schema-validator';
