+++ server/middlewares/validators/index.ts
@@ -1,0 +1,2 @@
+export * from './mongo-id-param-validator';
+export * from './zod-schema-validator';
